+++ vyre-macros/tests/ui/derive_laws_on_union.rs
@@ -1,9 +1,0 @@
-use vyre_macros::AlgebraicLaws;
-
-#[derive(AlgebraicLaws)]
-#[vyre(laws = [Commutative])]
-union BadLawTarget {
-    bits: u32,
-}
-
-fn main() {}